+++ editors/zed/languages/eu4/indents.scm
@@ -1,0 +1,7 @@
+(block "}" @end) @indent
+
+(parameter_block
+  (parameter_condition)
+  "]"
+  (_)*
+  "]" @end) @indent